Support final targets
This commit implements a game state and support on the aeroplane class to specify a final target.
This can be either a runway or a waypoint.

If specified the strip displays the target.
Statistics update according to if the aircraft reached its target.
If no target specified, all arrival landings and departure map exits are deemed successful.